services/embedding-worker/requirements.txt
aiokafka==0.11.0
qdrant-client==1.12.1
sentence-transformers==3.3.1
pydantic-settings==2.6.1
opentelemetry-sdk==1.29.0
opentelemetry-exporter-otlp-proto-grpc==1.29.0
